Remove global Builder
